feature: We now parse incoming json and emit docs.
The couchdb changes feed looks like:

{"results":[
{<change>}
,
{<change>}
,
[etc]

We now skip the prologue and also trim prepended commas. (Newline
delimited json is much nicer to parse, but sadly this isn't what we
get.)

Also, single chunks emitted by the underlying I/O layer might not be
complete json blobs. We attempt to parse each as json. On failure, we
store the incomplete chunk and move on. If we have a pre-existing chunk,
we prepend it to an incoming chunk and attempt to parse the result.

This approach is guaranteed to be buggy in that it won't handle the case
of a chunk that spans across two or parts of two distinct records. The
right approach is to split on the comma-newline delimiter and parse each piece.

Renamed `Package` to `Event` because this is a generic couchdb follower,
not an npm-registry-specific follower. Made its struct fields public so
they can be used.

Updated the example to match the new usage.

Additionally, bumped some dependencies. Hyper is now pulled from the
master branch, not from the tokio branch.